// Frontend/src/app/core/services/aminety-service/models/SlotResponseDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type SlotResponseDto = {
    id?: string;
    slotLabel?: string | null;
    slotDate?: string;
    startTime?: string;
    endTime?: string;
    maxCapacity?: number;
    currentBookings?: number;
};