// Frontend/src/app/core/services/aminety-service/models/CreateSlotRequestDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CreateSlotRequestDto = {
    slotLabel?: string | null;
    slotDate?: string;
    startTime?: string;
    endTime?: string;
    maxCapacity?: number;
};